- Fixes to type hashing.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/loading@36677 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/imap/resource.h
@@ -28,7 +28,7 @@ namespace CS
 {
   namespace Resource
   {
-    typedef size_t TypeID;
+    typedef int TypeID;
   }
 }
